Project 9/requirements.txt
pandas==2.1.3
numpy==1.24.3
matplotlib==3.8.2
ccxt==4.0.91
python-dotenv==1.0.0 